backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.6.0
pydantic[email]>=2.9.0
pyjwt[cryptography]>=2.9.0
passlib[bcrypt]>=1.7.4
httpx>=0.28.0
openai-agents>=0.0.7
openai-chatkit>=0.1.0
